refactor: update path at every file
